backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
replicate==0.34.1
Pillow==10.4.0
numpy==1.26.4
httpx==0.27.2
python-dotenv==1.0.1
